Demos: move colored-Checkbox example from Palette to Checkbox page
The Checkbox color prop is a Checkbox feature; its demo belongs on
the Checkbox page. PaletteDemo now focuses on the palette itself
(swatch grid + consumer-mapping pattern) without the Checkbox
example.

CheckboxDemo gains a "Color tagging" example with the same six
team-colored checkboxes that previously lived in PaletteDemo (width
locked at 280px so the layout doesn't shift as the "Selected: …"
text grows).
